+++ packages/components/src/utils/detect-native-popover-case.ts
@@ -1,0 +1,14 @@
+import { getHasNativePopoverSupport } from './';
+import { findClosestComponent } from './dom';
+
+export const detectNativePopoverCase = (host: HTMLElement, nested: boolean): boolean => {
+  if (getHasNativePopoverSupport()) {
+    const parentTableElement = findClosestComponent(
+      nested ? ((host.getRootNode() as ShadowRoot).host as HTMLElement) : host,
+      'pTable'
+    );
+    if (!!parentTableElement) {
+      return true;
+    }
+  }
+};